requirements.txt
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
rich==13.6.0
schema==0.7.5
urllib3==2.0.6
